Updated branding (#4)
* Update color palettes in tailwind and material styles

This commit changes the color values for 'primary' and introduces 'accent' in the color configuration for Tailwind CSS. Corresponding changes have been made in the material.scss file as well. These adjustments allow for a new range of colors to be utilized in the application UI.

* Add LogoIconComponent to Auth and Navbar modules

The LogoIconComponent has been added to the Auth and Navbar modules. It has been imported in the sign-up and sign-in modules and integrated into their respective component templates. Additionally, the navbar layout has been updated to include the LogoIconComponent.