Update navbar profile widget (#9)
* Replace hardcoded profile images with dynamic ones

The commit introduces dynamic generation of profile images in the navbar layout, instead of using hardcoded URLs. The 'profileImageUri$' function generates a URL for a profile image, based on the user's full name. An image URL is then generated for every user individually.

* Remove 'Your Profile' and 'Settings' options from navbar layout

The commit removes the 'Your Profile' and 'Settings' options from the navbar layout, reducing clutter on the page. This commit cleans up the navbar and makes the 'Sign out' option the lone item in the dropdown. This enhances the navbar layout by minimizing the user choices in the dropdown menu.